functions/requirements.txt
firebase_functions~=0.1.0
readability-lxml
lxml_html_clean
